Added some Authorization

+++ src/tasks/tasks.service.ts
@@ -5,6 +5,8 @@ import {TaskRepository} from "./task.repository";
 import {Task} from "./task.entity";
 import {TaskStatus} from "./task-status.enum";
 import {GetTasksFilterDto} from "./dto/get-tasks-filter.dto";
+import {User} from "../auth/user.entity";
+import {GetUser} from "../auth/get-user.decorator";
 
 @Injectable()
 export class TasksService {
@@ -15,12 +17,13 @@ constructor(
 }
 
 
-    async getTasks(filterDto: GetTasksFilterDto): Promise<Task[]>{
-     return  this.taskRepository.getTasks(filterDto)
+    async getTasks(filterDto: GetTasksFilterDto,
+                   user :User): Promise<Task[]>{
+     return  this.taskRepository.getTasks(filterDto,user)
     }
 
-    async getTaskById(id : number): Promise<Task> {
-        const  found = await this.taskRepository.findOne(id)
+    async getTaskById(id : number, user: User): Promise<Task> {
+        const  found = await this.taskRepository.findOne({where:{id, userId: user.id}})
         if(!found){
                     throw new NotFoundException()
                 }
@@ -28,12 +31,14 @@ constructor(
     }
 
 
-    async createTask(createTaskDTO: CreateTaskDto): Promise<Task> {
+    async createTask(createTaskDTO: CreateTaskDto,
+                     @GetUser()user : User
+                     ): Promise<Task> {
 
-        return  this.taskRepository.createTask(createTaskDTO);
+        return  this.taskRepository.createTask(createTaskDTO, user);
     }
-    async deleteTask(id : number): Promise<void>{
-        const  found = await this.taskRepository.delete(id)
+    async deleteTask(id : number, user:User): Promise<void>{
+        const  found = await this.taskRepository.delete({id,userId: user.id})
         if(found.affected === 0){
             throw new NotFoundException(`task with id ${id} not found `)
         }
@@ -42,8 +47,8 @@ constructor(
 
 
     }
-    async updateTaskStatus(id : number, status : TaskStatus): Promise<Task>{
-        const  task = await this.getTaskById(id)
+    async updateTaskStatus(id : number, status : TaskStatus,user:User): Promise<Task>{
+        const  task = await this.getTaskById(id,user)
         task.status = status
         await task.save()
         return task;
